backend/pyproject.toml
[project]
name = "spotifygraph"
version = "0.1.0"
description = ""
authors = [
    {name = "BARSOT Thomas",email = "thomas.barsot@insa-rouen.fr"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.133.0,<0.134.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
